// typings/sarif-schema.d.ts
export interface MultiformatMessageString {
  /**
   * A plain text message string or format string.
   */
  text: string;
  /**
   * A Markdown message string or format string.
   */
  markdown?: string;
  /**
   * Key/value pairs that provide additional information about the message.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
/**
 * Metadata that describes a specific report produced by the tool, as part of the analysis it provides or its runtime reporting.
 */
export interface ReportingDescriptor {
  /**
   * A stable, opaque identifier for the report.
   */
  id: string;
  /**
   * An array of stable, opaque identifiers by which this report was known in some previous version of the analysis tool.
   */
  deprecatedIds?: string[];
  /**
   * A unique identifer for the reporting descriptor in the form of a GUID.
   */
  guid?: string;
  /**
   * An array of unique identifies in the form of a GUID by which this report was known in some previous version of the analysis tool.
   */
  deprecatedGuids?: string[];
  /**
   * A report identifier that is understandable to an end user.
   */
  name?: string;
  /**
   * An array of readable identifiers by which this report was known in some previous version of the analysis tool.
   */
  deprecatedNames?: string[];
  /**
   * A message string or message format string rendered in multiple formats.
   */
  shortDescription?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A message string or message format string rendered in multiple formats.
   */
  fullDescription?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A set of name/value pairs with arbitrary names. Each value is a multiformatMessageString object, which holds message strings in plain text and (optionally) Markdown format. The strings can include placeholders, which can be used to construct a message in combination with an arbitrary number of additional string arguments.
   */
  messageStrings?: {
    [k: string]: MultiformatMessageString;
  };
  /**
   * Default reporting configuration information.
   */
  defaultConfiguration?: {
    /**
     * Specifies whether the report may be produced during the scan.
     */
    enabled?: boolean;
    /**
     * Specifies the failure level for the report.
     */
    level?: "none" | "note" | "warning" | "error";
    /**
     * Specifies the relative priority of the report. Used for analysis output only.
     */
    rank?: number;
    /**
     * Contains configuration information specific to a report.
     */
    parameters?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
    /**
     * Key/value pairs that provide additional information about the reporting configuration.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A URI where the primary documentation for the report can be found.
   */
  helpUri?: string;
  /**
   * A message string or message format string rendered in multiple formats.
   */
  help?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * An array of objects that describe relationships between this reporting descriptor and others.
   */
  relationships?: ReportingDescriptorRelationship[];
  /**
   * Key/value pairs that provide additional information about the report.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
/**
 * Information about the relation of one reporting descriptor to another.
 */
export interface ReportingDescriptorRelationship {
  /**
   * A reference to the related reporting descriptor.
   */
  target:
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      };
  /**
   * A set of distinct strings that categorize the relationship. Well-known kinds include 'canPrecede', 'canFollow', 'willPrecede', 'willFollow', 'superset', 'subset', 'equal', 'disjoint', 'relevant', and 'incomparable'.
   */
  kinds?: string[];
  /**
   * A description of the reporting descriptor relationship.
   */
  description?:
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      };
  /**
   * Key/value pairs that provide additional information about the reporting descriptor reference.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}